+++ src/api/dishes.ts
@@ -1,15 +1,5 @@
 import { db } from '@/config/firebaseClient'
 
-export interface Dish {
-  available: boolean
-  description: string
-  imageURL: string
-  name: string
-  preparationTime: number
-  price: number
-  servings: number
-}
-
 const dishCollection = db.collection('dishes')
 
 export async function createDish(dish: Dish) {
