+++ src/components/Admin/Layout/index.tsx
@@ -1,17 +1,14 @@
 import React, { FC } from 'react'
 import { Flex } from '@chakra-ui/layout'
 
-import AuthProvider from '@/components/Auth/AuthProvider'
-import Navbar from '@/components/Navbar'
+import Navbar from '@/components/admin/Navbar'
 
 const Layout: FC = ({ children }) => {
   return (
-    <AuthProvider>
-      <Flex h="100vh" w="100vw">
-        <Navbar />
-        {children}
-      </Flex>
-    </AuthProvider>
+    <Flex h="100vh" w="100vw">
+      <Navbar />
+      {children}
+    </Flex>
   )
 }
