--- src/components/client/MenuLayout/components/index.tsx
@@ -1,1 +1,0 @@
-export { default as Header } from './Header'
