+++ src/typings/index.d.ts
@@ -1,7 +1,10 @@
 interface Dish {
-  id: string
-  name: string
+  id?: string
+  available: boolean
   description: string
   imageURL: string
+  name: string
+  preparationTime: number
   price: number
+  servings: number
 }
